+++ apps/blog/admin.py
@@ -16,10 +16,7 @@
 """Representation of blog models in the admin interface."""
 
 from django import forms
-from django.contrib import admin, messages
-from django.db.models.query import QuerySet
-from django.template.defaultfilters import truncatechars
-from django.urls import reverse
+from django.contrib import admin
 from django.utils.html import format_html
 from django.utils.translation import gettext_lazy
 
@@ -132,103 +129,3 @@ class PostAdmin(BaseAdmin):
             )
         }),
     )
-
-
-@admin.register(Comment)
-class CommentAdmin(BaseAdmin):
-    """Class to manage post comments."""
-
-    list_display = (
-        'comment_sender',
-        'comment_content',
-        'post',
-        'object_status',
-        'created_at',
-    )
-
-    list_display_links = (
-        'comment_content',
-    )
-
-    list_filter = (
-        'status',
-        'created_at',
-        'updated_at',
-    )
-
-    search_fields = (
-        'user_name',
-        'user_email',
-        'comment',
-    )
-
-    fieldsets = (
-        (gettext_lazy('Commentator'), {
-            'fields': (
-                'user_name',
-                'user_email',
-            ),
-        }),
-        (gettext_lazy('Comment'), {
-            'fields': (
-                'comment',
-                'status',
-            )
-        }),
-    )
-
-    unpublished_statuses = (
-        Comment.STATUS_HIDDEN,
-    )
-
-    actions = (
-        'publish_comments',
-        'unpublish_comments',
-    )
-
-    @admin.action(description=gettext_lazy('Publish selected comments'))
-    def publish_comments(self, request, queryset: QuerySet):
-        """Publish selected comments."""
-        rows = queryset.update(status=Comment.STATUS_PUBLISHED)
-
-        if rows > 0:
-            self.message_user(
-                request,
-                gettext_lazy('Selected comments published successfully'),
-                messages.SUCCESS
-            )
-        else:
-            self.message_user(
-                request,
-                gettext_lazy('Nothing has been changed')
-            )
-
-    @admin.action(description=gettext_lazy('Unpublish selected comments'))
-    def unpublish_comments(self, request, queryset: QuerySet):
-        """Unpublish selected comments."""
-        rows = queryset.update(status=Comment.STATUS_HIDDEN)
-
-        if rows > 0:
-            self.message_user(
-                request,
-                gettext_lazy('Selected comments unpublished successfully'),
-                messages.SUCCESS
-            )
-        else:
-            self.message_user(
-                request,
-                gettext_lazy('Nothing has been changed')
-            )
-
-    @admin.display(ordering='user_name', description=gettext_lazy('Sender'))
-    def comment_sender(self, obj):
-        """Return custom column for comment sender."""
-        return format_html('<a href="{}">{}</a>'.format(
-            reverse("admin:blog_comment_change", args=(obj.pk,)),
-            (obj.user_name or obj.user_email)
-        ))
-
-    @admin.display(ordering='content', description=gettext_lazy('Comment'))
-    def comment_content(self, obj):
-        """Return custom column for comment body."""
-        return truncatechars(obj.comment.replace('\n', ''), 120)